requirements.txt
wandb>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jinja2>=3.1.0
requests>=2.31.0
scipy>=1.10.0
PyYAML>=6.0 